fix(stella-tools): chase the scratch_path signature through all exec callers
The constructor-injection rework widened drive/drive_split/run/run_github/
run_and_report with a trailing scratch_path but left 17 call sites and two
internal wrappers un-updated, and StartProcess's tuple-to-named-fields change
left one self.0 behind — the crate did not compile. Non-shell tools (ci,
github_rest, issue_ops, issues, verify, project, screenshot) pass None: the
scratch export surface is bash/process/scripts by construction.

scripts.rs had grown 12 lines past its closed god-file ceiling; its inline
tests module moves to scripts/tests.rs (the process/tests.rs pattern from the
sibling branch), taking the file to 1354 lines. No test bodies changed. Drops
the now-unused PathBuf import in subprocess_env.

Refs #2696
